backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
openai
pydantic
streamlit
requests
python-multipart
python-docx 
pdfplumber